[refactor][jinj00oo]: ImageList 역할 범위 및 파일 위치 수정 (#177)
- 이미지 상태 관리를 부모에서 관리
- ImageList는 UI 역할만 담당
