+++ src/entities/post/post-image/model/types.ts
