+++ src/entities/post/post-image/ui/ImageList.stories.tsx
@@ -1,10 +1,10 @@
 import type { Meta, StoryObj } from '@storybook/nextjs';
 import { useRef, useState } from 'react';
 import { ImageList } from './ImageList';
+import { reorderArray } from '../lib/reorder';
 
 const meta: Meta<typeof ImageList> = {
-  title: 'Features/UI/PostImage/ImageList',
-  component: ImageList,
+  title: 'Entities/UI/PostImage/ImageList',
   parameters: {
     layout: 'centered',
   },
@@ -21,14 +21,25 @@ export const Default: Story = {
     const inputRef = useRef<HTMLInputElement>(null);
     const [files, setFiles] = useState<File[]>([]);
 
+    /** 파일 선택 */
     const handleSelect = (e: React.ChangeEvent<HTMLInputElement>) => {
       if (!e.target.files) return;
       const selectedFiles = Array.from(e.target.files);
-      setFiles(selectedFiles);
+      setFiles((prev) => [...prev, ...selectedFiles]);
+    };
+
+    /** 파일 삭제 */
+    const handleRemove = (index: number) => {
+      setFiles((prev) => prev.filter((_, i) => i !== index));
+    };
+
+    /** 드래그 순서 변경 */
+    const handleReorder = (from: number, to: number) => {
+      setFiles((prev) => reorderArray(prev, from, to));
     };
 
     return (
-      <div className="flex w-[360px] flex-col items-center gap-4 border border-gray-200 p-4">
+      <div className="flex w-[360px] flex-col items-center gap-4 rounded-lg border border-gray-200 p-4">
         {/* 파일 업로드 버튼 */}
         <button
           type="button"
@@ -38,7 +49,7 @@ export const Default: Story = {
           이미지 추가
         </button>
 
-        {/* 숨겨진 input */}
+        {/* 숨겨진 파일 input */}
         <input
           ref={inputRef}
           type="file"
@@ -48,8 +59,8 @@ export const Default: Story = {
           onChange={handleSelect}
         />
 
-        {/* 파일 선택 시 즉시 반영 */}
-        <ImageList files={files} />
+        {/* 렌더링 전용 ImageList */}
+        <ImageList files={files} onRemove={handleRemove} onReorder={handleReorder} />
       </div>
     );
   },
